#2957: reviewed implementation and made some adjustments.
The "maximum lift of whole aircraft" is only described in the arXiv version of the paper but has been removed in the peer-reviewed version for ESWA. There seems to be an error in the pre-print (x5 and x6 stand for only one variable in the original formula and the ranges for valid values are missing).

I decided to keep the problem and assumed a rather large range to make the problem more interesting for benchmarking purposes. All differences to the published version of the problem instances are described.

git-svn-id: https://src.heuristiclab.com/svn/core/trunk@16394 2abd9481-f8db-48e9-bd25-06bc13291c1b